fix(dotnet): carry the whole sequence a system address holds
The field is 32 bits wide at the largest size class, and it was read into a
32-bit signed number. Every value at or above two to the thirty-first
wrapped negative, so a read answered a name no reader accepts: the address
ending in seven read as "Thob AA-A h-2147483648". About one address in
sixteen of the whole range is affected, and none of it reported an error.

The other direction lost the same half. A name stating a sequence above
what a signed number holds answered nothing, so half the legal range could
not be reached from a name at all.

The sequence is a long from the read to the name. The number before it
stays the boxel's own, which seven bits hold, and a name stating a larger
one names no boxel.

+++ dotnet/src/EliteDangerousAlmanac/Astronomy/ProceduralSystem.cs
@@ -73,7 +73,7 @@ public sealed class ProceduralSystem
     public char MassCode => MassCodes.FromSizeClass(parts.SizeClass);
 
     /// <summary>The system's own number inside its boxel.</summary>
-    public int Sequence => parts.N2;
+    public long Sequence => parts.N2;
 
     /// <summary>The parts the name is written from.</summary>
     /// <remarks>
